requirements.txt
Flask
google-genai
python-dotenv
markdown